Clean up an unreachable emptied channel closure
This resolves my concern about the duplicated logic: there was no sense
in duplicating this logic in the first place.

I think the idea of a "workState-level abstraction" is also out. As long
as there's a potential need to interact with the reattachers channel,
it's not meaningful to treat workState as separate from the broader Map.
I mainly separated it in the first place to clarify which fields stateMu
protects.