Use a channel to wait on tasks
There still seems to be an issue with sync.WaitGroup under synctest.
An even worse issue, in fact: the linked list in the heap span that
stores the bubble association for each WaitGroup seems to be getting
corrupted. It's supposed to be ordered by the offset of the object
within the span, but when I managed to catch this under a debugger
I noticed cases where I could follow the links and get lower offsets.

I have a pretty minimal reproducer, and will probably report it to the
Go issue tracker. However, this was always a pretty nonstandard use of
sync.WaitGroup, so for the time being I'm going to unblock myself.