Deal with a couple more mutation test findings
One of these I put a TODO on, because I really need to set this project
aside for a while.

One is that the `!detached && !task.result.Returned()` condition for
when we spawn a new worker can mutate to simply `!detached`. This is
correct, since we're handing off our work grant and never using it
again. It's just inefficient, so not something we can test directly.
This one's going to stick around, like not locking the state when we
have no keys to queue.